added detection and handling for SHELL_FAMILY == 'powershell', win_vmware_tools is no longer necessary
